+++ test/generators.spec.js
@@ -27,11 +27,27 @@ const OPTS = {
     commands: 'Commands',
     views: 'resources/views',
     migrations: 'database/migrations',
-    seeds: 'database/seeds'
+    seeds: 'database/seeds',
+    providers: 'providers'
   }
 }
 
 test.group('Generators', () => {
+  test('get path to the provider file', (assert) => {
+    const filePath = generators.provider.getFilePath('Event', OPTS)
+    assert.equal(filePath, path.join(__dirname, 'providers', 'EventProvider.js'))
+  })
+
+  test('make provider file singular', (assert) => {
+    const filePath = generators.provider.getFilePath('Events', OPTS)
+    assert.equal(filePath, path.join(__dirname, 'providers', 'EventProvider.js'))
+  })
+
+  test('normalize provider keyword', (assert) => {
+    const filePath = generators.provider.getFilePath('EventProvider', OPTS)
+    assert.equal(filePath, path.join(__dirname, 'providers', 'EventProvider.js'))
+  })
+
   test('get path to the controller file', (assert) => {
     const filePath = generators.httpController.getFilePath('User', OPTS)
     assert.equal(filePath, path.join(__dirname, 'app/Controllers/Http', 'UserController.js'))
@@ -49,7 +65,7 @@ test.group('Generators', () => {
 
   test('get data for the controller', (assert) => {
     const data = generators.httpController.getData('User', {})
-    assert.deepEqual(data, { name: 'UserController', resource: false })
+    assert.deepEqual(data, { name: 'UserController', resource: false, resourceName: 'user', resourceNamePlural: 'users' })
   })
 
   test('get path to the model file', (assert) => {